lgc last commit

+++ web10g-kis-0.13-5.4/net/ipv4/tcp_lgc.h
@@ -7,7 +7,7 @@
 extern const u32 lgc_log_lut[LGC_LUT_SIZE];
 extern const u32 lgc_exp_lut[LGC_LUT_SIZE];
 
-inline u32 lgc_log_lut_lookup(u16);
-inline u32 lgc_exp_lut_lookup(u16);
+inline u32 lgc_log_lut_lookup(u32);
+inline u32 lgc_exp_lut_lookup(u32);
 
 #endif
